Small refactoring #2
