+++ src/OpenHdWebUi.FileSystem/Consts.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenHdWebUi.FileSystem;
+
+public static class Consts
+{
+    /// <summary>
+    /// 0777 Permission
+    /// </summary>
+    public static UnixFileMode Mode0777 => (UnixFileMode)511;
+}
