+++ src/OpenHdWebUi.sln
@@ -14,6 +14,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		..\postinst = ..\postinst
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConversionDemo", "ConversionDemo\ConversionDemo.csproj", "{516B1232-6903-420E-9DD2-10187AFA5850}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenHdWebUi.FileSystem", "OpenHdWebUi.FileSystem\OpenHdWebUi.FileSystem.csproj", "{2F0BB713-E306-4A24-BA63-322A2A90EF84}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenHdWebUi.FFmpeg", "OpenHdWebUi.FFmpeg\OpenHdWebUi.FFmpeg.csproj", "{E8B63594-4E6E-479A-80ED-A2A401EF0452}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -28,6 +34,18 @@ Global
 		{76046B4D-733A-4C83-8C66-C661C23FCB31}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{76046B4D-733A-4C83-8C66-C661C23FCB31}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{76046B4D-733A-4C83-8C66-C661C23FCB31}.Release|Any CPU.Build.0 = Release|Any CPU
+		{516B1232-6903-420E-9DD2-10187AFA5850}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{516B1232-6903-420E-9DD2-10187AFA5850}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{516B1232-6903-420E-9DD2-10187AFA5850}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{516B1232-6903-420E-9DD2-10187AFA5850}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2F0BB713-E306-4A24-BA63-322A2A90EF84}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2F0BB713-E306-4A24-BA63-322A2A90EF84}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2F0BB713-E306-4A24-BA63-322A2A90EF84}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2F0BB713-E306-4A24-BA63-322A2A90EF84}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E8B63594-4E6E-479A-80ED-A2A401EF0452}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E8B63594-4E6E-479A-80ED-A2A401EF0452}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E8B63594-4E6E-479A-80ED-A2A401EF0452}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E8B63594-4E6E-479A-80ED-A2A401EF0452}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
